Extract current ParallelExecutor into PlatformThreadsParallelExecutor
