+++ src/test/java/org/abelsromero/parallels/jobs/ParallelExecutorTest.java
@@ -8,13 +8,13 @@ import java.util.concurrent.Callable;
 import static java.lang.Boolean.TRUE;
 import static org.assertj.core.api.Assertions.assertThat;
 
-class ParallelExecutorTest {
+abstract class ParallelExecutorTest {
 
     // aka. you only had one job
     @Test
     void should_run_a_single_job() {
         // when
-        final ParallelExecutor executor = new ParallelExecutor(1, 1);
+        final ParallelExecutor executor = ParallelExecutor.platformThreads(1, 1);
         final ExecutionDetails details = executor.run(() -> {
             Thread.sleep(10);
             return TRUE;
@@ -31,7 +31,7 @@ class ParallelExecutorTest {
         final int workers = 1;
         final int executions = 1000;
         // when
-        final ParallelExecutor executor = new ParallelExecutor(workers, executions);
+        final ParallelExecutor executor = ParallelExecutor.platformThreads(workers, executions);
         final ExecutionDetails details = executor.run(() -> {
             Thread.sleep(1);
             return TRUE;
@@ -48,7 +48,7 @@ class ParallelExecutorTest {
         final int workers = 8;
         final int executions = 1000;
         // when
-        final ParallelExecutor executor = new ParallelExecutor(workers, executions + 300);
+        final ParallelExecutor executor = ParallelExecutor.platformThreads(workers, executions + 300);
         final ExecutionDetails details = executor.run(new Callable<>() {
             private int counter = -1;
 
@@ -92,7 +92,7 @@ class ParallelExecutorTest {
         final int workers = 2;
         final int executions = 10;
         // when
-        final ParallelExecutor executor = new ParallelExecutor(workers, executions);
+        final ParallelExecutor executor = ParallelExecutor.platformThreads(workers, executions);
         final ExecutionDetails details = executor.run(() -> {
             throw new InterruptedException();
         });
