+++ src/test/java/org/abelsromero/parallels/jobs/PlatformThreadParallelExecutorTest.java
@@ -1,0 +1,6 @@
+package org.abelsromero.parallels.jobs;
+
+class PlatformThreadParallelExecutorTest extends ParallelExecutorTest {
+
+
+}
